requirements.txt
langchain
openai
chromadb
tiktoken
pinecone-client
